refactor: 2주차 코드 리뷰 반영

+++ src/main/java/com/example/demo/service/CustomerService.java
@@ -5,16 +5,14 @@ import com.example.demo.dto.CustomerDto;
 import com.example.demo.repository.customer.CustomerRepository;
 import java.util.List;
 import java.util.UUID;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
-
-    public CustomerService(CustomerRepository customerRepository) {
-        this.customerRepository = customerRepository;
-    }
 
     public CustomerDto save(String name, int age) {
         Customer customer = new Customer(name, age);
@@ -22,7 +20,7 @@ public class CustomerService {
         return CustomerDto.from(customer);
     }
 
-    public CustomerDto read(UUID id) {
+    public CustomerDto findById(UUID id) {
         return CustomerDto.from(customerRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("id에 해당하는 customer가 없습니다.")));
     }
@@ -36,10 +34,18 @@ public class CustomerService {
     }
 
     public void updateName(UUID id, String name) {
+        if (customerRepository.notExistById(id)) {
+            throw new IllegalArgumentException(String.format("고객 name 업데이트 오류 : 입력 된 id에 해당하는 고객이 존재하지 않습니다. 입력 받은 id : %s ", id.toString()));
+        }
+
         customerRepository.updateName(id, name);
     }
 
     public void deleteById(UUID id) {
+        if (customerRepository.notExistById(id)) {
+            throw new IllegalArgumentException(String.format("고객 삭제 오류 : 입력 된 id에 해당하는 고객이 존재하지 않습니다. 입력 받은 id : %s ", id.toString()));
+        }
+
         customerRepository.deleteById(id);
     }
 }
